requirements.txt
# ============================================================
# FRAUD DETECTION AGENT — requirements.txt
# Python 3.11 recommandé
# ============================================================

# --- Core Data Science ---
numpy==1.26.4
pandas==2.2.2
scipy==1.13.0

# --- Machine Learning ---
scikit-learn==1.4.2
xgboost==2.0.3
imbalanced-learn==0.12.2      # SMOTE
joblib==1.4.0

# --- Deep Learning (Autoencoder) ---
torch==2.3.0
torchvision==0.18.0

# --- Explainability (XAI) ---
shap==0.45.1
lime==0.2.0.1

# --- LLM / Agent ---
langchain==0.2.1
langchain-community==0.2.1    # Intégration Ollama local
langchain-core==0.2.1
ollama==0.2.1                  # Client Ollama Python

# --- API Backend ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-multipart==0.0.9

# --- Dashboard Frontend ---
streamlit==1.35.0
plotly==5.22.0
altair==5.3.0

# --- Visualisation (Notebooks & EDA) ---
matplotlib==3.9.0
seaborn==0.13.2

# --- Utilitaires ---
python-dotenv==1.0.1
loguru==0.7.2                  # Logging structuré
tqdm==4.66.4
pytest==8.2.0
httpx==0.27.0                  # Client HTTP pour tests FastAPI

# --- Notebooks ---
jupyter==1.0.0
ipykernel==6.29.4
